# MESMER-X: use params instead of distr