<!-- Feel free to remove check-list items aren't relevant to your change -->

 - [x] Closes #532
 - [ ] Tests added
 - [ ] Fully documented, including `CHANGELOG.rst`

This avoids `Expression.evaluate(...)` in favor of `Expression.evaluate_params(...)`, which leads to a approximately 2x-4x speedup of calibrating mesmer-x. Instead of adding many methods to `Expression` it calls the method on `Expression.distr` (e.g. `Expression.distr.pdf(x, **params)`. 

cc @l-pierini, @yquilcaille