@mathause 2025-02-03
Speedup


**py3.13 ubuntu-latest**

Before (6m 38s total)

```
169.00s call   tests/integration/test_calibrate_mesmer_x.py::test_calibrate_mesmer_x[ssp585-tasmax-norm(loc=c1 + c2 * __tas__, scale=c3)-expr1_2ndfit-True-False]
85.83s call    tests/integration/test_calibrate_mesmer_x.py::test_calibrate_mesmer_x[ssp585-tasmax-norm(loc=c1 + c2 * __tas__, scale=c3)-expr1-False-False]
8.70s call     tests/unit/test_mesmer_x_first_guess.py::test_first_guess_truncnorm
```

After (3m 43s total)

```
56.41s call    tests/integration/test_calibrate_mesmer_x.py::test_calibrate_mesmer_x[ssp585-tasmax-norm(loc=c1 + c2 * __tas__, scale=c3)-expr1_2ndfit-True-False]
29.68s call    tests/integration/test_calibrate_mesmer_x.py::test_calibrate_mesmer_x[ssp585-tasmax-norm(loc=c1 + c2 * __tas__, scale=c3)-expr1-False-False]
2.75s call     tests/unit/test_mesmer_x_first_guess.py::test_first_guess_truncnorm
```

---

**py3.13 windows-latest**


Before (11m 40s)

```
229.83s call     tests/integration/test_calibrate_mesmer_x.py::test_calibrate_mesmer_x[ssp585-tasmax-norm(loc=c1 + c2 * __tas__, scale=c3)-expr1_2ndfit-True-False]
116.79s call     tests/integration/test_calibrate_mesmer_x.py::test_calibrate_mesmer_x[ssp585-tasmax-norm(loc=c1 + c2 * __tas__, scale=c3)-expr1-False-False]
11.97s call     tests/unit/test_mesmer_x_first_guess.py::test_first_guess_truncnorm
```


After (9m 52s)
```

75.44s call     tests/integration/test_calibrate_mesmer_x.py::test_calibrate_mesmer_x[ssp585-tasmax-norm(loc=c1 + c2 * __tas__, scale=c3)-expr1_2ndfit-True-False]
39.01s call     tests/integration/test_calibrate_mesmer_x.py::test_calibrate_mesmer_x[ssp585-tasmax-norm(loc=c1 + c2 * __tas__, scale=c3)-expr1-False-False]
3.44s call     tests/unit/test_mesmer_x_first_guess.py::test_first_guess_truncnorm
```